refactor: separate tests mocks files
