+++ tests/domain/mocks/account/index.ts
@@ -1,1 +1,1 @@
-export * from './mock-account'
+export * from './mock-account-authentication'
