+++ tests/domain/mocks/category/mock-category-comands.ts
@@ -1,7 +1,6 @@
 import { faker } from '@faker-js/faker'
 
 import { type AddCategory, type UpdateCategory } from '@/domain/usecases/category/commands'
-import { type GetAllCategories } from '@/domain/usecases/category/queries'
 import { CategoryHelper } from '@/domain/entities/category'
 
 export const mockAddCategoryInput = (): AddCategory.Input => ({
@@ -9,16 +8,6 @@ export const mockAddCategoryInput = (): AddCategory.Input => ({
 })
 
 export const mockUpdateCategoryInput = (): UpdateCategory.Input => ({
-  id: faker.string.uuid(),
-  name: CategoryHelper.formatName(faker.string.alpha({ length: { min: 3, max: 20 } }))
-})
-
-export const mockCategoriesOutput = (): GetAllCategories.Output[] => [
-  mockCategoryOutput(),
-  mockCategoryOutput()
-]
-
-const mockCategoryOutput = (): GetAllCategories.Output => ({
   id: faker.string.uuid(),
   name: CategoryHelper.formatName(faker.string.alpha({ length: { min: 3, max: 20 } }))
 })
