+++ tests/presentation/mocks/category/mock-category-commands.ts
@@ -1,6 +1,4 @@
-import { mockCategoriesOutput } from '@/tests/domain/mocks/category'
 import { type AddCategory, type UpdateCategory } from '@/domain/usecases/category/commands'
-import { type GetAllCategories } from '@/domain/usecases/category/queries'
 
 export class AddCategorySpy implements AddCategory {
   public input: AddCategory.Input
@@ -15,13 +13,5 @@ export class UpdateCategorySpy implements UpdateCategory {
 
   public async update(input: UpdateCategory.Input): Promise<void> {
     this.input = input
-  }
-}
-
-export class GetAllCategoriesSpy implements GetAllCategories {
-  public output: GetAllCategories.Output[] = mockCategoriesOutput()
-
-  public async getAll(): Promise<GetAllCategories.Output[]> {
-    return this.output
   }
 }
